--- a/src/checkbox-group/checkbox-group.ts
+++ b/src/checkbox-group/checkbox-group.ts
@@ -6,7 +6,7 @@ const { prefix } = config;
 const name = `${prefix}-checkbox-group`;
 @wxComponent()
 export default class CheckBoxGroup extends SuperComponent {
-  externalClasses = [`${prefix}-class`];
+  externalClasses = ['t-class'];
 
   relations = {
     '../checkbox/checkbox': {
@@ -26,6 +26,7 @@ export default class CheckBoxGroup extends SuperComponent {
     prefix,
     classPrefix: name,
     checkboxOptions: [],
+    indeterminate: false,
   };
 
   properties = {
@@ -60,52 +61,42 @@ export default class CheckBoxGroup extends SuperComponent {
     getChilds() {
       let items = this.getRelationNodes('../checkbox/checkbox');
       if (!items.length) {
-        items = this.selectAllComponents('.t-checkbox-option');
+        items = this.selectAllComponents(`.${prefix}-checkbox-option`);
       }
       return items || [];
     },
 
-    // slot插入选项
     updateChildren() {
       const items = this.getChilds();
+      const { value, options } = this.data;
 
-      const { value, disabled } = this.data;
       if (items.length > 0) {
         items.forEach((item: any) => {
           !item.data.checkAll &&
             item.setData({
-              checked: value?.indexOf(item.data.value) > -1,
+              checked: value?.includes(item.data.value),
             });
-          item.setDisabled(disabled);
         });
         // 关联可全选项
-        if (items.findIndex((item) => item.data.checkAll) > -1) {
-          items.forEach((item) => {
-            item.setOptionLinked(true);
-          });
-          this.handleHalfCheck(items.length);
+        if (options.some((item) => item.checkAll)) {
+          this.handleCheckall();
         }
       }
     },
 
-    updateValue({ name, checked }) {
+    updateValue({ key, checked }) {
       const { value, max } = this.data;
       let newValue = value;
-      const keySet = this.getChilds()
-        .map((item) => item.data.value)
-        .reduce((set, key) => {
-          set.add(key);
-          return set;
-        }, new Set());
+      const keySet = new Set(this.getChilds().map((item) => item.data.value));
 
       newValue = newValue.filter((value) => keySet.has(value));
-      if (max && checked && newValue.length === max) {
-        return;
-      }
+
+      if (max && checked && newValue.length === max) return;
+
       if (checked) {
-        newValue = newValue.concat(name);
+        newValue = newValue.concat(key);
       } else {
-        const index = newValue.findIndex((v: string) => v === name);
+        const index = newValue.findIndex((v: string) => v === key);
         newValue.splice(index, 1);
       }
       this._trigger('change', { value: newValue });
@@ -114,9 +105,7 @@ export default class CheckBoxGroup extends SuperComponent {
     initWithOptions() {
       const { options } = this.data;
 
-      if (!options?.length || !Array.isArray(options)) {
-        return;
-      }
+      if (!options?.length || !Array.isArray(options)) return;
 
       const checkboxOptions = options.map((item) => {
         const isLabel = ['number', 'string'].includes(typeof item);
@@ -131,50 +120,44 @@ export default class CheckBoxGroup extends SuperComponent {
       this.setData({
         checkboxOptions,
       });
-
-      this.updateChildren(); // remove
     },
 
-    // 处理全选
-    handleCheckAll(e) {
-      const { checked, option, name } = e.detail || e;
-      const items = this.getChilds();
+    handleInnerChildChange(e) {
+      const { item } = e.target.dataset;
+      const { checked } = e.detail;
+      const { checkboxOptions, indeterminate } = this.data;
 
-      if (!option) {
-        if (!items?.length) {
-          return;
-        }
-        this._trigger('change', {
-          value: items
-            .map((item) => {
-              if (item.data.disabled) {
-                return this.data.value.includes(item.data.value) ? item.data.value : '';
-              }
-              return checked && !item.data.checkAll ? item.data.value : '';
-            })
-            .filter((val) => val),
-        });
+      if (item.checkAll) {
+        const value =
+          !checked && indeterminate
+            ? checkboxOptions.map((item) => item.value)
+            : checkboxOptions
+                .filter((item) => {
+                  if (item.disabled) {
+                    return this.data.value.includes(item.value);
+                  }
+                  return checked && !item.checkAll;
+                })
+                .map((item) => item.value);
+        this._trigger('change', { value });
       } else {
-        this.updateValue({ name, checked });
+        this.updateValue({ key: item.value, checked });
       }
     },
 
-    // 处理options半选
-    handleHalfCheck(len: number) {
-      const items = this.getChilds();
-      const checkboxOptions = items.filter((i) => !i.data.checkAll);
-      const all = checkboxOptions.map((item) => item.data.value);
-      const enableValue = checkboxOptions.filter((i) => !i.data.disabled).map((item) => item.data.value);
-      const currentVal = Array.from(new Set(this.data.value?.filter((i) => all.indexOf(i) > -1)));
-      const element = items.find((item) => item.data.checkAll);
-      if (currentVal.length) {
-        element?.setData({ checked: true });
-        element?.changeCheckAllHalfStatus(currentVal.length !== len - 1);
-        // 取消全选
-        element?.setCancel(enableValue.every((val) => currentVal.includes(val)));
-      } else {
-        element?.setData({ checked: false });
-      }
+    handleCheckall() {
+      const { checkboxOptions, value } = this.data;
+      const valueSet = new Set(value);
+      const isCheckall = checkboxOptions.every((item) => (item.checkAll ? true : valueSet.has(item.value)));
+      const items = this.selectAllComponents(`.${prefix}-checkbox-option`);
+      const $target = items.find((item) => item.data.checkAll);
+
+      $target.setData({
+        checked: valueSet.size > 0,
+      });
+      this.setData({
+        indeterminate: !isCheckall,
+      });
     },
   };
 }