--- a/src/checkbox-group/checkbox-group.wxml
+++ b/src/checkbox-group/checkbox-group.wxml
@@ -1,8 +1,8 @@
-<view class="{{ classPrefix }} t-class" style="{{customStyle}}">
+<view class="{{ classPrefix }} {{prefix}}-class" style="{{customStyle}}">
   <slot />
   <block wx:for="{{checkboxOptions}}" wx:key="value">
     <t-checkbox
-      class="t-checkbox-option"
+      class="{{prefix}}-checkbox-option"
       label="{{item.label || item.text || ''}}"
       value="{{item.value || ''}}"
       check-all="{{item.checkAll}}"