--- a/src/checkbox-group/checkbox-group.wxml
+++ b/src/checkbox-group/checkbox-group.wxml
@@ -6,8 +6,10 @@
       label="{{item.label || item.text || ''}}"
       value="{{item.value || ''}}"
       check-all="{{item.checkAll}}"
+      indeterminate="{{item.checkAll && indeterminate}}"
       disabled="{{item.disabled}}"
-      bind:toggleAll="handleCheckAll"
+      data-item="{{item}}"
+      bind:change="handleInnerChildChange"
     ></t-checkbox>
   </block>
 </view>