Checkbox: 重构代码，优化渲染性能

## 结果

**使用 checkbox 组件**

渲染个数|优化前耗时|优化后耗时|效果|
--|--|--|--
50|≈1000ms|≈110ms|+909%
100|≈9500ms|≈160ms|+5937%

**使用 options 属性**

渲染个数|优化前耗时|优化后耗时|效果
--|--|--|--
100|≈330ms|≈170ms|+193%
1000|≈2500ms|≈1100ms|+227%

> 测试设备：iPhone 12 ；系统版本 iOS 15.4；微信版本：8.0.20

resolve #352